+++ gson/src/test/java/feign/gson/GsonCodecTest.java
@@ -60,9 +60,12 @@ public class GsonCodecTest {
     Map<String, Object> map = new LinkedHashMap<String, Object>();
     map.put("foo", 1);
 
-    Response response =
-        Response.create(200, "OK", Collections.<String, Collection<String>>emptyMap(),
-                        "{\"foo\": 1}", UTF_8);
+    Response response = Response.builder()
+            .status(200)
+            .reason("OK")
+            .headers(Collections.<String, Collection<String>>emptyMap())
+            .body("{\"foo\": 1}", UTF_8)
+            .build();
     assertEquals(new GsonDecoder().decode(response, new TypeToken<Map<String, Object>>() {
     }.getType()), map);
   }
@@ -115,26 +118,34 @@ public class GsonCodecTest {
     zones.add(new Zone("denominator.io."));
     zones.add(new Zone("denominator.io.", "ABCD"));
 
-    Response response =
-        Response.create(200, "OK", Collections.<String, Collection<String>>emptyMap(), zonesJson,
-                        UTF_8);
+    Response response = Response.builder()
+            .status(200)
+            .reason("OK")
+            .headers(Collections.<String, Collection<String>>emptyMap())
+            .body(zonesJson, UTF_8)
+            .build();
     assertEquals(zones, new GsonDecoder().decode(response, new TypeToken<List<Zone>>() {
     }.getType()));
   }
 
   @Test
   public void nullBodyDecodesToNull() throws Exception {
-    Response response = Response.create(204, "OK",
-                                        Collections.<String, Collection<String>>emptyMap(),
-                                        (byte[]) null);
+    Response response = Response.builder()
+            .status(204)
+            .reason("OK")
+            .headers(Collections.<String, Collection<String>>emptyMap())
+            .build();
     assertNull(new GsonDecoder().decode(response, String.class));
   }
 
   @Test
   public void emptyBodyDecodesToNull() throws Exception {
-    Response response = Response.create(204, "OK",
-                                        Collections.<String, Collection<String>>emptyMap(),
-                                        new byte[0]);
+    Response response = Response.builder()
+            .status(204)
+            .reason("OK")
+            .headers(Collections.<String, Collection<String>>emptyMap())
+            .body(new byte[0])
+            .build();
     assertNull(new GsonDecoder().decode(response, String.class));
   }
 
@@ -181,8 +192,12 @@ public class GsonCodecTest {
     zones.add(new Zone("DENOMINATOR.IO.", "ABCD"));
 
     Response response =
-        Response.create(200, "OK", Collections.<String, Collection<String>>emptyMap(), zonesJson,
-                        UTF_8);
+        Response.builder()
+                .status(200)
+                .reason("OK")
+                .headers(Collections.<String, Collection<String>>emptyMap())
+                .body(zonesJson, UTF_8)
+                .build();
     assertEquals(zones, decoder.decode(response, new TypeToken<List<Zone>>() {
     }.getType()));
   }
@@ -214,9 +229,11 @@ public class GsonCodecTest {
   /** Enabled via {@link feign.Feign.Builder#decode404()} */
   @Test
   public void notFoundDecodesToEmpty() throws Exception {
-    Response response = Response.create(404, "NOT FOUND",
-        Collections.<String, Collection<String>>emptyMap(),
-        (byte[]) null);
+    Response response = Response.builder()
+            .status(404)
+            .reason("NOT FOUND")
+            .headers(Collections.<String, Collection<String>>emptyMap())
+            .build();
     assertThat((byte[]) new GsonDecoder().decode(response, byte[].class)).isEmpty();
   }
 }