+++ jackson/src/test/java/feign/jackson/JacksonCodecTest.java
@@ -80,27 +80,34 @@ public class JacksonCodecTest {
     zones.add(new Zone("denominator.io."));
     zones.add(new Zone("denominator.io.", "ABCD"));
 
-    Response response =
-        Response.create(200, "OK", Collections.<String, Collection<String>>emptyMap(), zonesJson,
-                        UTF_8);
+    Response response = Response.builder()
+                .status(200)
+                .reason("OK")
+                .headers(Collections.<String, Collection<String>>emptyMap())
+                .body(zonesJson, UTF_8)
+                .build();
     assertEquals(zones, new JacksonDecoder().decode(response, new TypeReference<List<Zone>>() {
     }.getType()));
   }
 
   @Test
   public void nullBodyDecodesToNull() throws Exception {
-    Response
-        response =
-        Response
-            .create(204, "OK", Collections.<String, Collection<String>>emptyMap(), (byte[]) null);
+    Response response = Response.builder()
+            .status(204)
+            .reason("OK")
+            .headers(Collections.<String, Collection<String>>emptyMap())
+            .build();
     assertNull(new JacksonDecoder().decode(response, String.class));
   }
 
   @Test
   public void emptyBodyDecodesToNull() throws Exception {
-    Response response = Response.create(204, "OK",
-                                        Collections.<String, Collection<String>>emptyMap(),
-                                        new byte[0]);
+    Response response = Response.builder()
+            .status(204)
+            .reason("OK")
+            .headers(Collections.<String, Collection<String>>emptyMap())
+            .body(new byte[0])
+            .build();
     assertNull(new JacksonDecoder().decode(response, String.class));
   }
 
@@ -114,9 +121,12 @@ public class JacksonCodecTest {
     zones.add(new Zone("DENOMINATOR.IO."));
     zones.add(new Zone("DENOMINATOR.IO.", "ABCD"));
 
-    Response response =
-        Response.create(200, "OK", Collections.<String, Collection<String>>emptyMap(), zonesJson,
-                        UTF_8);
+    Response response = Response.builder()
+            .status(200)
+            .reason("OK")
+            .headers(Collections.<String, Collection<String>>emptyMap())
+            .body(zonesJson, UTF_8)
+            .build();
     assertEquals(zones, decoder.decode(response, new TypeReference<List<Zone>>() {
     }.getType()));
   }
@@ -205,9 +215,11 @@ public class JacksonCodecTest {
   /** Enabled via {@link feign.Feign.Builder#decode404()} */
   @Test
   public void notFoundDecodesToEmpty() throws Exception {
-    Response response = Response.create(404, "NOT FOUND",
-        Collections.<String, Collection<String>>emptyMap(),
-        (byte[]) null);
+    Response response = Response.builder()
+            .status(404)
+            .reason("NOT FOUND")
+            .headers(Collections.<String, Collection<String>>emptyMap())
+            .build();
     assertThat((byte[]) new JacksonDecoder().decode(response, byte[].class)).isEmpty();
   }
 }