+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### Version 9.2
+* Supports context path when using Ribbon `LoadBalancingTarget`
+* Adds builder methods for the Response object
+* Deprecates Response factory methods
+* Adds nullable Request field to the Response object
+
+### Version 9.1
+* Allows query parameters to match on a substring. Ex `q=body:{body}`
+
 ### Version 9.0
 * Migrates to maven from gradle
 * Changes maven groupId to `io.github.openfeign`